TASK: Move tree state into separate files
