+++ Resources/Private/JavaScript/asset-collections/src/hooks/useSelectedAssetCollection.ts
@@ -2,7 +2,7 @@ import { useRecoilValue } from 'recoil';
 import { useQuery } from '@apollo/client';
 
 import AssetCollection from '../interfaces/AssetCollection';
-import selectedAssetCollectionIdState from '../state/selectedAssetCollectionIdState';
+import { selectedAssetCollectionIdState } from '../state/selectedAssetCollectionIdState';
 import ASSET_COLLECTION from '../queries/assetCollection';
 
 interface AssetCollectionQueryResult {
