+++ Resources/Private/JavaScript/asset-collections/src/state/assetCollectionFocusedState.ts
@@ -1,0 +1,12 @@
+import { selectorFamily } from 'recoil';
+
+import { selectedAssetCollectionIdState } from './selectedAssetCollectionIdState';
+
+// This state selector provides the focused state for each individual asset collection
+export const assetCollectionFocusedState = selectorFamily<boolean, string>({
+    key: 'AssetCollectionFocusedState',
+    get:
+        (assetCollectionId) =>
+        ({ get }) =>
+            get(selectedAssetCollectionIdState) === assetCollectionId,
+});
