+++ Resources/Private/JavaScript/asset-collections/src/state/createAssetCollectionDialogVisibleState.ts
@@ -1,8 +1,6 @@
 import { atom } from 'recoil';
 
-const createAssetCollectionDialogVisibleState = atom<boolean>({
+export const createAssetCollectionDialogVisibleState = atom<boolean>({
     key: 'createAssetCollectionDialogState',
     default: false,
 });
-
-export default createAssetCollectionDialogVisibleState;
