+++ Resources/Private/JavaScript/asset-collections/src/state/selectedAssetCollectionIdState.ts
@@ -1,10 +1,8 @@
 import { atom } from 'recoil';
 import { localStorageEffect } from '@media-ui/core/src/state';
 
-const selectedAssetCollectionIdState = atom<string | null>({
+export const selectedAssetCollectionIdState = atom<string | null>({
     key: 'SelectedAssetCollectionIdState',
     default: null,
     effects: [localStorageEffect('SelectedAssetCollectionIdState')],
 });
-
-export default selectedAssetCollectionIdState;
